refactor all models and controllers create api for vacaciones,vacacionesestado,vacacionestipo

+++ api/database/migrations/2020_07_26_030710_create_vacaciones_tipo_table.php
@@ -16,14 +16,14 @@ class CreateVacacionesTipoTable extends Migration
         Schema::create('vacaciones_tipo', function (Blueprint $table) {
             $table->id();
             $table->string('tipo');
-            $table->integer('tiempo_dias');
+            $table->decimal('tiempo_dias');
             $table->timestamps();
             $table->softDeletes();
         });
 
         Schema::create('vacaciones_tipo_history', function (Blueprint $table) {
             $table->string('tipo');
-            $table->integer('tiempo_dias');
+            $table->decimal('tiempo_dias');
             //control de historial
             $table->unsignedInteger('id');
             $table->timestamp('registerUtc');
